% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_yll_gformula_estimand_wrappers}
\alias{estimate_yll_gformula_estimand_wrappers}
\alias{estimate_yll_gformula_ate}
\alias{estimate_yll_gformula_att}
\alias{estimate_yll_gformula_atc}
\title{Estimand-specific wrappers}
\usage{
estimate_yll_gformula_ate(...)

estimate_yll_gformula_att(...)

estimate_yll_gformula_atc(...)
}
\arguments{
\item{...}{Arguments forwarded to \code{\link[=estimate_yll_gformula]{estimate_yll_gformula()}}.}
}
\value{
Same list structure as \code{\link[=estimate_yll_gformula]{estimate_yll_gformula()}}.
}
\description{
Thin wrappers around \code{\link[=estimate_yll_gformula]{estimate_yll_gformula()}} that pre-fill the
\code{estimand} argument.
}
