% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YLLgmethods-package.R
\docType{package}
\name{YLLgmethods-package}
\alias{YLLgmethods}
\alias{YLLgmethods-package}
\title{YLLgmethods: Years of Life Lost via the G-Formula}
\description{
Estimates Years of Life Lost (YLL) and counterfactual life expectancy
under user-specified interventions on a binary exposure using the
parametric g-formula on the attained-age time scale. Supports ATE, ATT,
ATC, and stochastic / population interventions, with bootstrap confidence
intervals and helpers for plotting conditional and marginal survival
curves.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShuntaroS/yll_g-methods}
  \item Report bugs at \url{https://github.com/ShuntaroS/yll_g-methods/issues}
}

}
\author{
\strong{Maintainer}: Shuntaro Sato \email{shuntaro.sato.stat@gmail.com} (\href{https://orcid.org/0000-0002-6055-3649}{ORCID})

}
\keyword{internal}
